third_party/blink/web_tests/gamepad/gamepad-detached-no-crash-expected.txt

Accessing Navigator Gamepad methods on a closed window.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS w.navigator.getGamepads().length is 4
PASS successfullyParsed is true

TEST COMPLETE
